fix(checkout): CHECKOUT-7606 remove snapshot tests for google autocomplete
