+++ packages/core/src/app/address/googleAutocomplete/GoogleAutocomplete.test.tsx
@@ -1,33 +1,33 @@
-import { render } from '@testing-library/react';
-import React from 'react';
+import '@testing-library/jest-dom';
+import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import React, { FunctionComponent } from 'react';
 
-import GoogleAutocomplete from './GoogleAutocomplete';
+import GoogleAutocomplete, { GoogleAutocompleteProps } from './GoogleAutocomplete';
 
 describe('GoogleAutocomplete Component', () => {
-    // @todo: add more tests when we can successfully mock "GoogleAutocompleteService"
-    // At the moment, jest.mock('...') doesn't seem to do the trick.
 
-    it('renders input with initial value', () => {
-        expect(
-            render(
-                <GoogleAutocomplete
-                    apiKey="bar"
-                    initialValue="fo"
-                    inputProps={{ placeholder: 'NO PO BOX' }}
-                />,
-            ),
-        ).toMatchSnapshot();
+    let ComponentTest: FunctionComponent<GoogleAutocompleteProps>;
+    let defaultProps: GoogleAutocompleteProps;
+
+    beforeAll(() => {
+        defaultProps = {
+            apiKey: "bar",
+            initialValue: "fo"
+        };
+        ComponentTest = (props) => (<GoogleAutocomplete {...props} />);
     });
 
-    it('renders input with max length', () => {
-        expect(
-            render(
-                <GoogleAutocomplete
-                    apiKey="bar"
-                    initialValue="fo"
-                    inputProps={{ maxLength: "50" }}
-                />,
-            ),
-        ).toMatchSnapshot();
+    it('renders input with initial value and placeholder text', () => {
+        render(<ComponentTest {...defaultProps} inputProps={{ placeholder: 'NO PO BOX' }} />);
+        expect(screen.getByPlaceholderText('NO PO BOX')).toBeInTheDocument();
+    });
+
+    it('renders input and limits value to max length', async () => {
+        render(<ComponentTest {...defaultProps} initialValue="" inputProps={{ placeholder: 'Max length 50', maxLength: "50" }} />);
+        expect(screen.getByPlaceholderText('Max length 50')).toBeInTheDocument();
+        await userEvent.type(screen.getByPlaceholderText('Max length 50'), '120 South Jean Baptiste Point du Sable Lake Shore Drive');
+        const inputElement: HTMLInputElement = screen.getByPlaceholderText('Max length 50');
+        expect(inputElement.value).toBe('120 South Jean Baptiste Point du Sable Lake Shore ');
     });
 });
